requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==3.2.13
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
Pillow==8.4.0
PyPDF2==1.27.12
python-bidi==0.4.2
pytz==2022.1
reportlab==3.5.53
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
webencodings==0.5.1
xhtml2pdf==0.2.5
